bug fix
fixed bug with the teleportation to the final battle
fixed bug with the creating the new observer when the hero is reincarnationg with the dafault ability.
now the observer will not be created at the team battle
fixed murloc (blue murloc)
fixed master necromancer heal due the Taste of death ability
fixed spawning of the two satyr barracks.